Move expand tests to a separate folder
